refactor useGetBrandingPreferenceResolve
